Cleaning up directories
Removed Org and mail Directory as they are not part of the Multiscad.  Org was the default xampp website, mail was a folder that held contact_me.php which was a trial file to text sending emails.  This may be re-implimented in the future, but for now it is not needed.