Fixed a JS scroll issue
Fixed a small bug in the code that caused the layout to be mis-aligned when scrolling from the bottom of the website to the top using the SVG input button.  This fix allowes a slight overshoot without altering how the website looks, it also makes the top navbar snap to the top 5 pixels sooner in mobile mode, and 1 pixel sooner in desktop mode.