Added Check codes
Added some codes to the menu script to account for the website being at the very top when selecting a link on a mobile device.  When at the very top, the menu pushes down the entire site, but after that the menu goes over the site.  The JS code now detects this and compensates the scroll amount to account for the menu when it is opened and your at the very top of the site.