Renamed css
Renamed component.css to input.css
Going to start splitting css files into different files to make managing it easier.
On the actual site, the css files will be combined into a single css file, and be minified tosave bandwidth, allowing the css to stay managable easily.
Will also be doing the same with the JS files as well.