Updated index.php
changed readfile(INPUTFILE) to require(INPUTFILE) so that the file loaded is treated as a php file and is processed as such.  Even though HTML files are loaded, they are still processed though PHP processor and anything inside php tags are computed as php.
This allows for a dynamic site without having to require the entire site be php coding.