Updated css loader
I updated the CSS loader to make it load css files dynamically instead of being hardcoded.  They are set up to be loaded in order based on there filename.  They are pre-named with a number for there order (0-9).
If more then 10 files are present, Use naming of 00-99 to prevent issues.