Adjusted layout system
Instead of the old scrolling layout, the site now utalizes a tab type system.  You click the link and it hides the old section and reveal the clicked section.
It can be reverted back to the old way by just removing 2_main.js and renaming 2_main.js.old to 2_main.js