Added Page Disabling
Added codes to header.html, index.php, and main.js to check for disabled pages.
Disabled pages will not be loaded in the body, and the link above will be grayed out and un-clickable.
This was added encase a page breaks the entire site, it can be disabled individually while the rest of the site remains online.

Use "<?php //DISABLED ?>" to disable a page in the body.