Lower Navbar Overrides
Lowered the Navbar over-rides from 1010px to 950px, allowing for the screen to shrink 60px in width more before collapsing the navigation bar to the dropdown menu.
This allows for screens with resolutions of 1024×768 to still see the full size navigation bar.  At 1010px, the navbar whould be in dropdown mode due to the browser border, as well as the possibility of a scrollbar being shown, cuasing the navbar width to be below 1010px.  as 950px, there is a 74px allowance before going into dropdown mode, which also means a scrollbar canstill be visable with a full size navbar at 1024x768 resolution.