Fixed JS bug
Fixed a JS bug caused by removing href from links.
Bug details: When loading the page, the bug prevented the JS code to mark the first link in the navbar as active, making it seem as if no link was active, when in fact, the first link is always active on page load.  This nos now fixed, and upon loading the page, the first link is now properly marked as active.